pyproject.toml
[tool.poetry]
name = "Produtitas"
version = "0.1.0"
description = "Microserviço responsável pelo gerenciamento de produto de dados"
authors = ["anapalopes <ana.lopes1@creditas.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.1.3"
Flask-Classful = "^0.14.2"
Flask-Cors = "^3.0.10"
Werkzeug = "2.1.2"
tomli = "^2.0.1"
requests = "^2.28.1"
python-dotenv = "^0.20.0"
kafka-python = "^2.0.2"
celery = "^5.2.7"
redis = "^4.3.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
